requirements.txt
Flask[async]
requests
twilio
deepgram-sdk
openai
aiohttp
python-dotenv
vocode